feat: cleaned up old js payment related components
