+++ frontend/web/components/ProjectsPage.tsx
@@ -3,7 +3,7 @@ import ProjectManageWidget from './ProjectManageWidget'
 import OrganisationProvider from 'common/providers/OrganisationProvider'
 import ConfigProvider from 'common/providers/ConfigProvider'
 import Project from 'common/project'
-import { onPaymentLoad } from './modals/Payment'
+import { onPaymentLoad } from './modals/payment/onPaymentLoad'
 import makeAsyncScriptLoader from 'react-async-script'
 import { useRouteMatch } from 'react-router-dom'
 import Utils from 'common/utils/utils'
